services/ml/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
redis==5.0.1
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
prophet==1.1.4
scipy==1.11.4
joblib==1.3.2